Separate checkRole test in a file
